Move categories fetch to vuex
